+++ jhotdraw-core/src/main/java/org/jhotdraw/draw/action/edit/BringToFrontAction.java
@@ -1,46 +1,44 @@
 /*
- * @(#)SendToBackAction.java
+ * @(#)BringToFrontAction.java
  *
  * Copyright (c) 2003-2008 The authors and contributors of JHotDraw.
  * You may not use, copy or modify this file, except in compliance with the
  * accompanying license terms.
  */
-package org.jhotdraw.draw.action;
+package org.jhotdraw.draw.action.edit;
 
-import dk.sdu.mmmi.featuretracer.lib.FeatureEntryPoint;
+import org.jhotdraw.draw.action.AbstractSelectedAction;
 import org.jhotdraw.draw.figure.Figure;
 import java.util.*;
 import javax.swing.undo.*;
 import org.jhotdraw.draw.*;
 import org.jhotdraw.util.ResourceBundleUtil;
+import dk.sdu.mmmi.featuretracer.lib.FeatureEntryPoint;
 
 /**
- * SendToBackAction.
+ * ToFrontAction.
  *
  * @author Werner Randelshofer
  * @version $Id$
  */
-public class SendToBackAction extends AbstractSelectedAction {
 
+public class BringToFrontAction extends AbstractSelectedAction {
     private static final long serialVersionUID = 1L;
-    public static final String ID = "edit.sendToBack";
+    public static final String ID = "edit.bringToFront";
 
-    /**
-     * Creates a new instance.
-     */
-    public SendToBackAction(DrawingEditor editor) {
+    public BringToFrontAction(DrawingEditor editor) {
         super(editor);
-        ResourceBundleUtil labels
-                = ResourceBundleUtil.getBundle("org.jhotdraw.draw.Labels");
-        labels.configureAction(this, ID);
-        updateEnabledState();
+        ResourceBundleUtil
+                .getBundle("org.jhotdraw.draw.Labels")
+                .configureAction(this, ID);
+        super.updateEnabledState();
     }
 
     @Override
     public void actionPerformed(java.awt.event.ActionEvent e) {
         final DrawingView view = getView();
         final LinkedList<Figure> figures = new LinkedList<>(view.getSelectedFigures());
-        sendToBack(view, figures);
+        bringToFront(view, figures);
         fireUndoableEditHappened(new AbstractUndoableEdit() {
             private static final long serialVersionUID = 1L;
 
@@ -54,22 +52,21 @@ public class SendToBackAction extends AbstractSelectedAction {
             @Override
             public void redo() throws CannotRedoException {
                 super.redo();
-                SendToBackAction.sendToBack(view, figures);
+                BringToFrontAction.bringToFront(view, figures);
             }
 
             @Override
             public void undo() throws CannotUndoException {
                 super.undo();
-                BringToFrontAction.bringToFront(view, figures);
+                SendToBackAction.sendToBack(view, figures);
             }
         });
     }
-
-    @FeatureEntryPoint("SendToBack feature")
-    public static void sendToBack(DrawingView view, Collection<Figure> figures) {
+    @FeatureEntryPoint("bringToFront feature")
+    public static void bringToFront(DrawingView view, Collection<Figure> figures) {
         Drawing drawing = view.getDrawing();
-        for (Figure figure : figures) { // XXX Shouldn't the figures be sorted here back to front?
-            drawing.sendToBack(figure);
+        for (Figure figure : drawing.sort(figures)) {
+            drawing.bringToFront(figure);
         }
     }
 }
