+++ jhotdraw-core/src/main/java/org/jhotdraw/draw/action/edit/SendToBackAction.java
@@ -1,43 +1,47 @@
 /*
- * @(#)BringToFrontAction.java
+ * @(#)SendToBackAction.java
  *
  * Copyright (c) 2003-2008 The authors and contributors of JHotDraw.
  * You may not use, copy or modify this file, except in compliance with the
  * accompanying license terms.
  */
-package org.jhotdraw.draw.action;
+package org.jhotdraw.draw.action.edit;
 
+import dk.sdu.mmmi.featuretracer.lib.FeatureEntryPoint;
+import org.jhotdraw.draw.action.AbstractSelectedAction;
 import org.jhotdraw.draw.figure.Figure;
 import java.util.*;
 import javax.swing.undo.*;
 import org.jhotdraw.draw.*;
 import org.jhotdraw.util.ResourceBundleUtil;
-import dk.sdu.mmmi.featuretracer.lib.FeatureEntryPoint;
 
 /**
- * ToFrontAction.
+ * SendToBackAction.
  *
  * @author Werner Randelshofer
  * @version $Id$
  */
+public class SendToBackAction extends AbstractSelectedAction {
 
-public class BringToFrontAction extends AbstractSelectedAction {
     private static final long serialVersionUID = 1L;
-    public static final String ID = "edit.bringToFront";
+    public static final String ID = "edit.sendToBack";
 
-    public BringToFrontAction(DrawingEditor editor) {
+    /**
+     * Creates a new instance.
+     */
+    public SendToBackAction(DrawingEditor editor) {
         super(editor);
-        ResourceBundleUtil
-                .getBundle("org.jhotdraw.draw.Labels")
-                .configureAction(this, ID);
-        super.updateEnabledState();
+        ResourceBundleUtil labels
+                = ResourceBundleUtil.getBundle("org.jhotdraw.draw.Labels");
+        labels.configureAction(this, ID);
+        updateEnabledState();
     }
 
     @Override
     public void actionPerformed(java.awt.event.ActionEvent e) {
         final DrawingView view = getView();
         final LinkedList<Figure> figures = new LinkedList<>(view.getSelectedFigures());
-        bringToFront(view, figures);
+        sendToBack(view, figures);
         fireUndoableEditHappened(new AbstractUndoableEdit() {
             private static final long serialVersionUID = 1L;
 
@@ -51,21 +55,22 @@ public class BringToFrontAction extends AbstractSelectedAction {
             @Override
             public void redo() throws CannotRedoException {
                 super.redo();
-                BringToFrontAction.bringToFront(view, figures);
+                SendToBackAction.sendToBack(view, figures);
             }
 
             @Override
             public void undo() throws CannotUndoException {
                 super.undo();
-                SendToBackAction.sendToBack(view, figures);
+                BringToFrontAction.bringToFront(view, figures);
             }
         });
     }
-    @FeatureEntryPoint("bringToFront feature")
-    public static void bringToFront(DrawingView view, Collection<Figure> figures) {
+
+    @FeatureEntryPoint("SendToBack feature")
+    public static void sendToBack(DrawingView view, Collection<Figure> figures) {
         Drawing drawing = view.getDrawing();
-        for (Figure figure : drawing.sort(figures)) {
-            drawing.bringToFront(figure);
+        for (Figure figure : figures) { // XXX Shouldn't the figures be sorted here back to front?
+            drawing.sendToBack(figure);
         }
     }
 }
