+++ jhotdraw-gui/src/main/java/org/jhotdraw/gui/action/ButtonFactory.java
@@ -71,7 +71,7 @@ import org.jhotdraw.draw.action.AlignAction;
 import org.jhotdraw.draw.action.ApplyAttributesAction;
 import org.jhotdraw.draw.action.AttributeAction;
 import org.jhotdraw.draw.action.AttributeToggler;
-import org.jhotdraw.draw.action.BringToFrontAction;
+import org.jhotdraw.draw.action.edit.BringToFrontAction;
 import org.jhotdraw.draw.action.ColorIcon;
 import org.jhotdraw.draw.action.DefaultAttributeAction;
 import org.jhotdraw.draw.action.DrawingAttributeAction;
@@ -88,7 +88,7 @@ import org.jhotdraw.draw.action.SelectSameAction;
 import org.jhotdraw.draw.action.SelectionColorChooserAction;
 import org.jhotdraw.draw.action.SelectionColorChooserHandler;
 import org.jhotdraw.draw.action.SelectionColorIcon;
-import org.jhotdraw.draw.action.SendToBackAction;
+import org.jhotdraw.draw.action.edit.SendToBackAction;
 import org.jhotdraw.draw.action.StrokeIcon;
 import org.jhotdraw.draw.action.UngroupAction;
 import org.jhotdraw.draw.action.ZoomAction;
